Move tinycthread into separate directory
