+++ .gitignore
@@ -2,7 +2,7 @@
 .Rhistory
 .RData
 .Ruserdata
-src/*.o
-src/*.so
-src/*.dll
+*.o
+*.so
+*.dll
 inst/doc
